+++ plugins/main/common/dashboards/dashboard-definitions/management/cluster/configuration/dashboard.ts
@@ -1,0 +1,36 @@
+import {
+  DashboardConfig,
+  DashboardLayoutDefinition,
+} from '../../../../lib/dashboard-config-service';
+import { getVisStateTop5Nodes } from './vis-states';
+
+export class ClusterConfigurationDashboardLayoutDefinition extends DashboardLayoutDefinition {
+  constructor(indexPatternId: string) {
+    super();
+    this.setGridVisualizationPairs({
+      gridData: { w: 48, h: 13, x: 0, y: 0 },
+      savedVis: getVisStateTop5Nodes(indexPatternId),
+    });
+  }
+}
+
+export class ClusterConfigurationDashboardConfig extends DashboardConfig {
+  constructor(indexPatternId: string) {
+    super(
+      indexPatternId,
+      new ClusterConfigurationDashboardLayoutDefinition(indexPatternId),
+    );
+  }
+
+  protected override getId(): string {
+    return 'cluster-configuration-dashboard-tab';
+  }
+
+  protected override getTitle(): string {
+    return 'Cluster configuration dashboard';
+  }
+
+  protected override getDescription(): string {
+    return 'Dashboard of the Cluster configuration';
+  }
+}