+++ plugins/main/server/health-check/index.ts
@@ -2,3 +2,4 @@ export * from './index-patterns';
 export * from './templates';
 export * from './server-api';
 export * from './types';
+export * from './saved-objects';