+++ plugins/main/server/plugin.ts
@@ -37,6 +37,7 @@ import {
   initializationTaskCreatorServerAPIConnectionCompatibility,
   mapFieldsFormat,
 } from './health-check';
+import { initializationTaskCreatorSavedObjectsForDashboardsAndVisualizations } from './health-check';
 import {
   FIELD_TIMESTAMP,
   HEALTH_CHECK_TASK_INDEX_PATTERN_AGENTS_MONITORING,
@@ -261,6 +262,10 @@ export class WazuhPlugin implements Plugin<WazuhPluginSetup, WazuhPluginStart> {
         },
         configurationSettingKey: 'pattern',
       }),
+    );
+
+    core.healthCheck.register(
+      initializationTaskCreatorSavedObjectsForDashboardsAndVisualizations(),
     );
 
     core.healthCheck.register(